[MetaxGPU][Tests] Fold PR33 regression checks into stable suites
